donor-main/databricks/requirements.txt
azure-storage-blob==12.22.0
faiss-cpu==1.8.0
langchain==0.2.15
langchain-openai==0.1.23
langchain-community==0.2.14
pymupdf==1.24.9
pdfminer.six==20231228
pdfplumber==0.11.4
PyPDF2==3.0.1
sqlalchemy==2.0.19
typing_extensions==4.12.2